Finished re-factoring Onpub frontend includes in to the new OnpubFrontend class.
No end-user changes are required as a results of this re-factoring.
This enables the new OnpubFrontend class to be sub-classed and extended
by an object called OnpubFrontendCustom. If this class file and class
are defined in the Onpub local/ dir, it will be loaded automatically
and its display method will be called.